+++ src/main/Main.java
@@ -6,6 +6,7 @@ import org.openqa.selenium.chrome.ChromeDriver;
 import methodTest.AccountUpdateTest;
 import methodTest.LoginTest;
 import methodTest.AccountInsertTest;
+import methodTest.UniformListTest;
 
 public class Main {
 
@@ -22,6 +23,7 @@ public class Main {
 		LoginTest logintest = new LoginTest(driver);
 		AccountInsertTest accountInsertTest = new AccountInsertTest(driver);
 		AccountUpdateTest accountUpdateTest = new AccountUpdateTest(driver);
+		UniformListTest uniformListTest = new UniformListTest(driver);
 
 		// 新規会員登録
 		Thread.sleep(500);
@@ -30,6 +32,10 @@ public class Main {
 		// ログイン
 		Thread.sleep(500);
 		logintest.Login();
+
+		// 商品一覧
+		Thread.sleep(500);
+		uniformListTest.UniformList();
 
 		//ユーザー情報変更
 		Thread.sleep(500);
